# perf: add thread local LRU caches for compiled programs