--- a/.github/workflows/eunit.yaml
+++ b/.github/workflows/eunit.yaml
@@ -34,6 +34,7 @@ jobs:
           brew update
           brew install erlang@${{ matrix.otp }}
           brew install automake
+          brew install bison
       - name: install rebar3
         run: |
           wget https://s3.amazonaws.com/rebar3/rebar3 && chmod +x rebar3