This PR adds thread local LRU (Least Recently Used) caches for caching `jq_state`s (holding the compiled versions of JQ filter programs). This greatly improves performance when the same filter program is used frequently. This is true when compiling the filter program is much slower than processing the JSON text with a compiled program (which seems to be very common). It is important the LRU cache is thread local since the jq_state struct instances are changed while they are used to process a JSON text.

The following table shows the throughput (processed `jq:parse/2` calls per second) reported by the test case `concurrent_queries_test_` before and after this commit on an Intel i5-1145G7 (16GB ram, 4 cores with hyper-threading):


```
N. of processes    Without LRU*     With LRU**
1                  100              290150
2                  200              422110
3                  270              625580
4                  360              731850
5                  330              725010
6                  420              716090
7                  410              758030
8                  250              777240

* Without LRU = processed `jq:parse/2` calls per second before this commit

** With LRU = processed `jq:parse/2` calls per second after this commit
```