@qzhuyan 2022-04-04
A general question is I see jq_state struct has pointers in it, I guess you have checked that the data they point to are safe to get reused since the memcpy will copy the pointer as well.