@qzhuyan 2022-04-05
> > A general question is I see jq_state struct has pointers in it, I guess you have checked that the data they point to are safe to get reused since the memcpy will copy the pointer as well.
> 
> As far as I have seen, the same jq_state struct can be reused to process multiple JSON objects. This is also tested in the concurrent_queries_test_ case (that reuses the same jq_state object many times due to the LRU cache). I don't think the jq_state struct is ever `memcpy'ed (if it was, it might be a problem). Please explain if I have misunderstood your concern.

After offline discussion, We have checked it is covered by unittest and memory sanitizer .